requirements.txt
flask
torch
torchvision
torchaudio
Pillow
trimesh
diffusers
accelerate
transformers
einops
xformers
numpy>=1.26.4,<2.0.0
scikit-learn
tqdm
rembg
onnxruntime-gpu
huggingface_hub
opencv-python-headless
tiktoken
sentencepiece
protobuf
